// packages/proto/src/wallet.pb.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.6
//   protoc               v3.21.12
// source: wallet.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { wrappers } from "protobufjs";
import { Observable } from "rxjs";
import { Struct } from "./google/protobuf/struct.pb";

export const protobufPackage = "wallet";

export interface CreateWalletRequest {
  userId: string;
}

export interface GetWalletRequest {
  userId: string;
}

export interface UpdateWalletBalanceRequest {
  userId: string;
  amount: number;
}

export interface CommonResponseArray {
  status?: number | undefined;
  success?: boolean | undefined;
  message: string;
  data: { [key: string]: any }[];
  errors?: string | undefined;
}

export interface CommonResponseObj {
  status?: number | undefined;
  success?: boolean | undefined;
  message: string;
  data?: { [key: string]: any } | undefined;
  errors?: string | undefined;
}

export const WALLET_PACKAGE_NAME = "wallet";

wrappers[".google.protobuf.Struct"] = { fromObject: Struct.wrap, toObject: Struct.unwrap } as any;

export interface WalletServiceClient {
  createWallet(request: CreateWalletRequest): Observable<CommonResponseObj>;

  getWallet(request: GetWalletRequest): Observable<CommonResponseObj>;

  creditWallet(request: UpdateWalletBalanceRequest): Observable<CommonResponseObj>;

  debitWallet(request: UpdateWalletBalanceRequest): Observable<CommonResponseObj>;
}

export interface WalletServiceController {
  createWallet(
    request: CreateWalletRequest,
  ): Promise<CommonResponseObj> | Observable<CommonResponseObj> | CommonResponseObj;

  getWallet(request: GetWalletRequest): Promise<CommonResponseObj> | Observable<CommonResponseObj> | CommonResponseObj;

  creditWallet(
    request: UpdateWalletBalanceRequest,
  ): Promise<CommonResponseObj> | Observable<CommonResponseObj> | CommonResponseObj;

  debitWallet(
    request: UpdateWalletBalanceRequest,
  ): Promise<CommonResponseObj> | Observable<CommonResponseObj> | CommonResponseObj;
}

export function WalletServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createWallet", "getWallet", "creditWallet", "debitWallet"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("WalletService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("WalletService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const WALLET_SERVICE_NAME = "WalletService";